Test 'non-nullable realm object reference'
