+++ generator/test/generator_test_io/non_nullable_ro_reference.log
@@ -1,0 +1,13 @@
+Realm object references must be nullable
+
+in: asset:pkg/test/generator_test_io/non_nullable_ro_reference.dart:13:8
+    ╷
+8   │ @RealmModel()
+9   │ class _Bad {
+    │       ━━━━ in realm model for 'Bad'
+... │
+13  │   late _Other wrong;
+    │        ^^^^^^ is not nullable
+    ╵
+Change type to _Other?
+
